Implementa endpoints de tipo de ingresso e tipo de cliente.

+++ src/main/java/br/com/tasima/ida/daprazuar/eventman/services/EventoService.java
@@ -69,7 +69,7 @@ public class EventoService {
 		eventoRepository.save(ev);
 	}
 
-	public Evento update(int id, Evento ev) throws InvalidParameterException, NotFoundException {
+	public Evento update(long id, Evento ev) throws InvalidParameterException, NotFoundException {
 		if (ev == null) {
 			throw new InvalidParameterException();
 		}
@@ -83,4 +83,5 @@ public class EventoService {
 
 		return localEv;
 	}
+	
 }
